[refactor] 롬복 및 패키지 구조 변경
